requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
protobuf==3.17.3
pyocr==0.8.1
python-dotenv==0.19.2
pywatchman==1.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.6
Werkzeug==2.0.1